+++ src/syssense/resources/__init__.py
@@ -1,0 +1,1 @@
+"""Recursos estáticos empacotados do SysSense."""